chore: refactor react dir
