+++ packages/client/src/react/components/Modals/Leaderboard.tsx
@@ -4,7 +4,7 @@ import styled from "styled-components";
 import { world } from "../../../mud/world";
 import { useMUD } from "../../../MUDContext";
 import { colors } from "../../styles/global";
-import { PlayerData, ShipData } from "../Modal";
+import { PlayerData, ShipData } from "./Modal";
 
 export function Leaderboard() {
   const {
