+++ packages/client/src/react/components/ShipStatus/RevealButtons.tsx
@@ -1,6 +1,6 @@
 import { useComponentValue } from "@latticexyz/react";
-import { useMUD } from "../../MUDContext";
-import { ConfirmButton, Success } from "../styles/global";
+import { useMUD } from "../../../MUDContext";
+import { ConfirmButton, Success } from "../../styles/global";
 
 export function RevealButtons({ acted }: { acted: boolean }) {
   const {
