+++ packages/client/src/react/components/ShipStatus/ShipAttribute.tsx
