+++ packages/client/src/react/components/ShipStatus/ShipCard.tsx
