+++ packages/client/src/react/components/ShipStatus/ShipDamage.tsx
