+++ packages/client/src/react/components/ShipStatus/YourShips.tsx
@@ -2,13 +2,13 @@ import { useComponentValue, useObservableValue } from "@latticexyz/react";
 import { EntityIndex, getComponentValue, Has, HasValue, runQuery, setComponent } from "@latticexyz/recs";
 import { merge } from "rxjs";
 import styled from "styled-components";
-import { world } from "../../mud/world";
-import { useMUD } from "../../MUDContext";
-import { usePlayer } from "../../PlayerContext";
-import { ModalType, Phase } from "../../types";
-import { ConfirmButtons } from "../ConfirmButtons";
-import { Button, colors, Container } from "../styles/global";
-import { Cell } from "./Cell";
+import { world } from "../../../mud/world";
+import { useMUD } from "../../../MUDContext";
+import { usePlayer } from "../../../PlayerContext";
+import { ModalType, Phase } from "../../../types";
+import { Button, colors, Container } from "../../styles/global";
+import { Cell } from "./../Cell";
+import { ConfirmButtons } from "./ConfirmButtons";
 import { YourShip } from "./YourShip";
 
 const gridConfig = { gridRowStart: 9, gridRowEnd: 13, gridColumnStart: 1, gridColumnEnd: 13 };
