+++ packages/client/src/react/components/TopBar.tsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import { useMUD } from "../../MUDContext";
 import { ShipAttributeTypes } from "../types";
 import { Cell } from "./Cell";
-import ShipAttribute from "./OverviewComponents/ShipAttribute";
+import ShipAttribute from "./ShipStatus/ShipAttribute";
 
 const gridConfig = { gridRowStart: 1, gridRowEnd: 3, gridColumnStart: 1, gridColumnEnd: 5 };
